userservice

+++ src/main/kotlin/servis/chat/response/FormResponse.kt
@@ -1,5 +1,5 @@
 package servis.chat.response
 
 class FormResponse (
-    val id:Int
+    val id:Long
 )
